Parse list and scalar args for torch ops (#16)
Resolves an issue with the `FxImporter` where arguments that are
`!torch.list` types are not properly imported due to the Fx tracer
denoting them via `immutable_list` types rather than encapsulating them
in an `fx.Node`. Also addresses a case where a python scalar is not
parsed into an `fx.Node` but rather retains its python type. For
instance, in
[`aten.convolution`](https://github.com/llvm/torch-mlir/blob/0109bf705b0dae3b4d55c7a9b180516787cbeac6/include/torch-mlir/Dialect/Torch/IR/GeneratedTorchOps.td#L4583C18-L4583C18)
where the "transposed" parameter is passed as false by default - this
argument is left as a `bool` and therefore needs to be handled when
importing arguments.