Document known issues (#29)
This includes a markdown file with documentation for some known issues.
The three included here comprise the vast majority of all errors that we
encounter when attempting to import novel `nn.Module` instances. Each
one also includes a minimal reproducible example. The vast majority of
other errors (at least in the tests that I have sampled) have to do with
unimplemented ops which can either be dealt with via including
decompositions or implementing ops in upstream torch-mlir.

Note: I included a small tweak to the importer, adding the ability to
convert `None` to the appropriate `!torch.none` in our `TypeSubclassMap`
because it 1) obscures the real issue in one of these cases and 2) is
probably something we want there anyways.

---------

Co-authored-by: brucekimrokcmu <kwangkyk@alumni.cmu.edu>