Add more decompositions (#34)
This addresses a slew of "unimplemented" errors in the test suite for
any operations that has an available decomposition. The only exception
is `aten.smooth_l1_loss` which, despite having a decomposition available
in pytorch, is somehow still triggering the same error - we can
investigate this later or implement the op if needed.