Address lift fresh copy (#37)
This PR implements a workaround to the issues we face as a result of
lift_fresh_copy. This op is generated to guarantee the freshness
invariant of a new tensor (see
[here](https://github.com/pytorch/pytorch/blob/8c62f01cb7b75cb647882a9bf85f10031b08896e/torch/fx/experimental/proxy_tensor.py#L329)).
FX actually takes constant tensor objects from operations like
`torch.Tensor([2.])` and [creates uniquely named
attributes](https://github.com/pytorch/pytorch/blob/36399d067a56d9875fd9c2bf61434126398ffb87/torch/fx/graph.py#L843C56-L843C56)
in the top level module to store their values, hence these values never
touch Turbine's importer (this is also why there is an apparent graph
break at these operations, they're referencing something that wasn't
created in the user-defined function). Dynamo then represents these
constant values with a `get_attr` node that seeks to retrieve the object
associated with that name from the top level module.

The fix implemented here is simple: catch when this occurs, grab the
actual object from the top level module, and create an entry in our
importer's `_v` table so that we can properly associate the identifier
generated by FX with the appropriate `MlirValue`. To do this we
construct a `torch.vtensor.literal` op which requires an `ElementsAttr`
as input. `ElementsAttr` requires a python buffer to create the
corresponding mlir `tensor` object. However, there is an issue: torch
`Tensor` objects do not currently implement the full python array
interface so they can't be directly converted into a python buffer (see
this [issue](https://github.com/pytorch/pytorch/issues/51156) and [this
article](https://pearu.github.io/array_interface_pytorch.html)). So we
have to convert the tensor to numpy in order to get a buffer version of
our tensor.

It's possible I overcomplicated this last part, and it has the
unfortunate side effect that we can't do this for `bfloat16` (because
this type doesn't exist in numpy) or `torch.complex` datatypes (because
the numpy buffer format for these objects is incompatible with the
buffer format `DenseElementsAttr` is expecting for complex datatypes -
so if there's a better way to do this I'd love to implement it that way
instead.

Generated tests are now at 90% passing.