Implement AOT argument specification and procedural tracing.
This now lets us have more natural compiled modules:

```
        class ProcArgsModule(CompiledModule):
            def foobar(self, a=AbstractTensor(3, 2), b=AbstractTensor(1, 1)):
                x = self.compute(a, b)
                y = self.compute(x, a)
                return y

            @jittable
            def compute(a, b):
                return a + b
```

In this example, the `foobar` function is an IREE procedural trace that can do anything with the IR. Right now, it supports calling functions and handling tensor metadata properly. The `compute` function is `@jittable` which says that we should use Dynamo and torch-mlir to compile this for use on IREE. Right now, we are taking it "all the way" to IREE internal IR, but I will add a mode later where we can stop at Torch IR and do the calls/casts properly.

After this, I need to thread through the globals/parameter handling, and then we will have an API for doing everything that the old module importer did (and more) but in a more flexible way, in pure python, on PyTorch2 (no TorchScript!).

Example output:

```
module @proc_args {
  func.func @foobar(%arg0: tensor<3x2xf32>, %arg1: tensor<1x1xf32>) -> tensor<3x2xf32> {
    %0 = call @compute(%arg0, %arg1) : (tensor<3x2xf32>, tensor<1x1xf32>) -> tensor<3x2xf32>
    %1 = call @compute$1(%0, %arg0) : (tensor<3x2xf32>, tensor<3x2xf32>) -> tensor<3x2xf32>
    return %1 : tensor<3x2xf32>
  }
  func.func private @compute(%arg0: tensor<3x2xf32>, %arg1: tensor<1x1xf32>) -> tensor<3x2xf32> {
    %0 = tensor.empty() : tensor<3x2xf32>
    %1 = linalg.generic {indexing_maps = [#map, #map1, #map], iterator_types = ["parallel", "parallel"]} ins(%arg0, %arg1 : tensor<3x2xf32>, tensor<1x1xf32>) outs(%0 : tensor<3x2xf32>) {
    ^bb0(%in: f32, %in_0: f32, %out: f32):
      %2 = arith.addf %in, %in_0 : f32
      linalg.yield %2 : f32
    } -> tensor<3x2xf32>
    return %1 : tensor<3x2xf32>
  }
  func.func private @compute$1(%arg0: tensor<3x2xf32>, %arg1: tensor<3x2xf32>) -> tensor<3x2xf32> {
    %0 = tensor.empty() : tensor<3x2xf32>
    %1 = linalg.generic {indexing_maps = [#map, #map, #map], iterator_types = ["parallel", "parallel"]} ins(%arg0, %arg1 : tensor<3x2xf32>, tensor<3x2xf32>) outs(%0 : tensor<3x2xf32>) {
    ^bb0(%in: f32, %in_0: f32, %out: f32):
      %2 = arith.addf %in, %in_0 : f32
      linalg.yield %2 : f32
    } -> tensor<3x2xf32>
    return %1 : tensor<3x2xf32>
  }
}
```