Test mnist training (#61)
Added MNIST training test case.
This requires torch-vision, so added it to the requirement. 

Furthermore, the change in cpu pass to `torch-to-iree`, which needs the
change in IREE pass to be landed and reflected in the IREE release, will
fail all the tests with `empty_strided` involved. As soon as IREE's new
release is available, the xfails can be removed.