Implement import phasing in AOT to avoid incremental compilation of j… (#72)
…ittables.

Primarily, this reworks jittable so that it always does a pure import
into the torch dialect, not lowering each compute function to linalg
incrementally. The result is that we have a module at the torch level
which is legal input to IREE's `torch-to-iree` input pipeline (vs the
result of that). In addition to likely being more resilient to change
(since we control the dialects), it should also be more debuggable since
it is easy now to dump to a file and continue with CL tools before any
real compilation has taken place.

In order to make this work, we add an `import_to=` kwarg to the
CompileModule that allows disabling any import-based compilation. Also
adds a couple of other APIs for ergonomics:

* `run_import`
* `run_pass_pipeline`
* `save_mlir`

By default, we run the full import pipeline as part of constructing the
module. We may change this in the future.