[RFC] basic model builder infra (#156)
The main motivation here is to be able to programmatically create hf
models (for SHARK) using the bare minimum information. And to have a
real api instead of importing from examples. I'm not sure if there are
any other use cases besides hf. Also sorely lacking in support for
anything that doesn't just work with aot.export, multiple inputs, etc.
Ideally this will also be the point at which we do any torch ir caching
cloudside.