[tk] Implement stream kernel code generation and hook to vector (#234)
This is still very much a work in progress as it skates over a number of
tricky issues to address when doing code emission to an actual
dispatch/stream executable. Specifically, this is as far as I can go
without building out symbolic expression support, since that is what is
needed for the workgroup calculations. This is also skating by on
dynamic shapes for the moment, although most of the plumbing is in place
to do it right.