Initial work on turbine-llm.
This is intended to be a new light-weight layer and model repository for constructing and converting models from various LLM toolchains for compilation via Turbine and serving via Turbine-Serving.

It is based on some earlier prototype work for producing PyTorch models that can run and compile directly off of GGUF files.