Update arg mapper in main kernel loop to map by stage
I noticed that some mmas were improperly using
the most recent values instead of iter args. In order to
fix this, I had to modify the mapper to only use iter args
and values within a node's stage. I also modified
a lot of the dicts to use the node as the key instead
of the node's name as this was also leading to errors during
printing.